Merge commit 'remotes/linuxport' into my-external-libraries-support

git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/external-libraries-support@18752 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/GUIWindowVideoBase.cpp
@@ -908,7 +908,7 @@ bool CGUIWindowVideoBase::OnClick(int iItem)
 {
   CFileItemPtr pItem = m_vecItems->Get(iItem);
 
-  if (!pItem->Exists())
+  if (!pItem->m_bIsFolder && !pItem->Exists())
   {
     if (!CGUIWindowVideoNav::DeleteItem(pItem.get(),true))
       return true;
